extract buildAndRunTestRunner + ensure it does not remove our container logs
