+++ testnet/e2e_v2_happy_path_test.go
@@ -1,13 +1,9 @@
 package main
 
 import (
-	"fmt"
 	"testing"
-	"time"
 
 	"github.com/ory/dockertest/v3"
-	"github.com/ory/dockertest/v3/docker"
-	"github.com/stretchr/testify/require"
 )
 
 func TestV2HappyPath(t *testing.T) {
@@ -16,43 +12,6 @@ func TestV2HappyPath(t *testing.T) {
 		pool *dockertest.Pool,
 		network *dockertest.Network,
 	) {
-		// bring up the test runner
-		t.Log("building and deploying test runner")
-		testRunner, err := pool.BuildAndRunWithBuildOptions(
-			&dockertest.BuildOptions{
-				Dockerfile: "testnet.Dockerfile",
-				ContextDir: "./orchestrator",
-			},
-			&dockertest.RunOptions{
-				Name:      "test_runner",
-				NetworkID: network.Network.ID,
-				PortBindings: map[docker.Port][]docker.PortBinding{
-					"8545/tcp": {{HostIP: "", HostPort: "8545"}},
-				},
-				Mounts: []string{
-					fmt.Sprintf("%s/testdata:/testdata", wd),
-				},
-				Env: []string{
-					"RUST_BACKTRACE=1",
-					"RUST_LOG=INFO",
-					"TEST_TYPE=V2_HAPPY_PATH",
-				},
-			},
-			func(config *docker.HostConfig) {},
-		)
-		require.NoError(t, err, "error bringing up test runner")
-
-		t.Logf("deployed test runner at %s", testRunner.Container.ID)
-		defer func() {
-			testRunner.Close()
-		}()
-
-		container := testRunner.Container
-		for container.State.Running {
-			time.Sleep(10 * time.Second)
-			container, err = pool.Client.InspectContainer(testRunner.Container.ID)
-			require.NoError(t, err, "error inspecting test runner")
-		}
-		require.Equal(t, 0, container.State.ExitCode, "container exited with error")
+		buildAndRunTestRunner(t, wd, pool, network, "V2_HAPPY_PATH")
 	})
 }
